+++ .dialyzer_ignore.exs
@@ -1,8 +1,8 @@
 [
   # Ignore warnings caused by the following problem:
-  # https://github.com/elixir-lang/elixir/issues/7177
+  # https://github.com/elixir-lang/elixir/issues/6738
   {"lib/xema/schema.ex", :pattern_match, 323},
-  {"lib/xema/validator.ex", :pattern_match, 388},
+  {"lib/xema/validator.ex", :pattern_match, 381},
 
   # due a bug in dialyzer
   ~r/.*Function.Xema.Castable.*.__impl__.1.does.not.exist.*/